Use Case
This is a simple use case with 3 manufacturers selling on Amazon. 

In order to run the network (assuming you setup your environment for HLF to run), follow the commands on "bash-commands" file. 

Also there are two versions of the app, Java and Go. Check Java folder and pcxchg for Go.